requirements.txt
dnspython>=2.4.0
python-whois>=0.8.0
aiohttp>=3.9.0
asyncio>=3.4.3
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
